+++ proyecto c++/FactoryIngresos.h
@@ -7,5 +7,5 @@ class FactoryIngresos {
 private:
 public:
 
-	static Ingresos* crearIngresos(int tipo, Colaborador* colab, int valor1 = 0, float valor2 = 0.0f);
+	static Ingresos* crearIngresos(int tipo, Colaborador* colab, int valor1 = 0, double valor2 = 0.0f);
 };
